INCOMPAT, exclusively use `mo` as suffix for months, `m` as suffix for minutes, fixes issue #76
We hereby break our promise to to disambiguate the `m` suffix when
adding durations to date/times based on the input.

This will break lines from the time when date and date/time tools were
separate.

Previously:

    $ dateadd 2018-02-02 +1m
    =>
      2018-03-02

Becomes:

    $ dateadd 2018-02-02 +1mo
    =>
      2018-03-02

Previously:

    $ dateseq 2010-01 +1m 2011-12

Becomes:

    $ dateseq 2010-01 +1mo 2011-12

The added-value of this incompatible change is that adding durations
to mixed date and date/time input is now unambiguous, e.g.

    $ dateadd +2mo <<EOF
    2018-02-02
    2018-02-02T10:00
    EOF
    =>
      2018-04-02
      2018-04-02T10:00:00

    $ dateadd +2m <<EOF
    2018-02-02
    2018-02-02T10:00
    EOF
    =>
      2018-02-02
      2018-02-02T10:02:00

    $ dateadd +2mo +2m <<EOF
    2018-02-02
    2018-02-02T10:00
    EOF
    =>
      2018-04-02
      2018-04-02T10:02:00